LPS-146271 Move container custom styles to general panel
